+++ js/package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "JSONStream": "^1.3.5",
     "cors": "^2.8.5",
     "cross-fetch": "^3.1.4",
     "d3": "^6.7.0",
